+++ src/elements/DEM_ellip3d/ellip3d/particle.h
@@ -1,6 +1,7 @@
 #ifndef PARTICLE_H
 #define PARTICLE_H
 
+#include "realtypes.h"
 #include "vec.h"
 #include "gradation.h"
 #include "contact.h"
@@ -17,20 +18,20 @@ class particle{
 	friend class contact<particle>;
 	friend class flb_bdry<particle>;
 public:
-	particle(int n, int type, vec center, long double r);
-	particle(int n, int type, vec center, long double a, long double b, long double c);
+	particle(int n, int type, vec center, REAL r);
+	particle(int n, int type, vec center, REAL a, REAL b, REAL c);
 	particle(int id,int type, vec dim, vec position, vec dirca, vec dircb, vec dircc);
 	particle(int n, int type, vec center, gradation& grad);
 
 	int    getID() const;
 	int    getType() const;
-	long double getA() const;
-	long double getB() const;
-	long double getC() const;
-	long double getRadius(vec v) const;
-	long double getVolume() const;
-	long double getMass() const;
-	long double getDensity() const;
+	REAL getA() const;
+	REAL getB() const;
+	REAL getC() const;
+	REAL getRadius(vec v) const;
+	REAL getVolume() const;
+	REAL getMass() const;
+	REAL getDensity() const;
 	vec    getJ() const;
 	vec    getCurrPosition() const;
 	vec    getPrevPosition() const;
@@ -52,17 +53,17 @@ public:
 	vec    getMoment() const;
 	vec    getConstForce() const;
 	vec    getConstMoment() const;
-	long double getTransEnergy() const;
-	long double getRotatEnergy() const;
-	long double getKinetEnergy() const;
-	long double getPotenEnergy(long double ref) const;
+	REAL getTransEnergy() const;
+	REAL getRotatEnergy() const;
+	REAL getKinetEnergy() const;
+	REAL getPotenEnergy(REAL ref) const;
 
 	void   setID(int n);
 	void   setType(int n);
-	void   setA(long double dd);
-	void   setB(long double dd);
-	void   setC(long double dd);
-	void   setMass(long double d);
+	void   setA(REAL dd);
+	void   setB(REAL dd);
+	void   setC(REAL dd);
+	void   setMass(REAL d);
 	void   setJ(vec vv);
 	void   setCurrPosition(vec vv);
 	void   setPrevPosition(vec vv);
@@ -80,7 +81,7 @@ public:
 	void   setPrevAcceleration(vec vv);
 	void   setCurrAlf(vec vv);
 	void   setPrevAlf(vec vv);
-	void   setDensity(long double dn) {density=dn;}
+	void   setDensity(REAL dn) {density=dn;}
 	void   setExternForce(vec fc) {force=fc;}
 	void   setExternMoment(vec mm) {moment=mm;}
 	void   setForce(vec vv);
@@ -96,8 +97,8 @@ public:
 	// update global coefficients in the following form based on position/dimensions/orientations
 	// a0 x^2 + a1 y^2 + a2 z^2 + a3 xy + a4 yz + a5 zx + a6 x + a7 y + a8 z + a9 = 0
 	void   GlobCoef();  
-	void   getGlobCoef(long double coef[]) const; // fetch global coeffs into coef[]
-	long double surfaceError(vec pt) const;
+	void   getGlobCoef(REAL coef[]) const; // fetch global coeffs into coef[]
+	REAL surfaceError(vec pt) const;
 	vec    localVec(vec) const;   // transform a vector in global coordinates into local coordinates
 	vec    globalVec(vec) const;  // transform a vector in local coordinates into global coordinates
 	void   print() const;
@@ -111,14 +112,14 @@ public:
 
 	//find the point on plane which is deepest into a particles, px+qy+rz+s=0 is the equation of the plane
 	//true means intersection; false means no intersection.
-	bool nearestPTOnPlane(long double p, long double q, long double r, long double s, vec& ptnp) const;
+	bool nearestPTOnPlane(REAL p, REAL q, REAL r, REAL s, vec& ptnp) const;
 
 	//side indicates which side the particles are in about the plane
 	//calculate the normal force between particle and a plane rigid boundary
 	void planeRBForce(plnrgd_bdry<particle>* plb,
 			  std::map<int,std::vector<boundarytgt> >& BoundarytgtMap,
 			  std::vector<boundarytgt>& vtmp,
-			  long double &penetr);
+			  REAL &penetr);
 
 	//if side<0, particles are inside the cylinder; else side is outside the cylinder
 	//calculate the normal force between particle and a cylinder wall
@@ -132,7 +133,7 @@ private:
         // 3-special case 2 (ellipsoidal pile): translation in vertical direction only
 	int ID;
 	int type;            
-	long double a,b,c;   // three semi-axles, a>=b>=c
+	REAL a,b,c;   // three semi-axles, a>=b>=c
 	vec curr_position;   // particle center
 	vec prev_position;
 	vec curr_direction_a, curr_direction_b, curr_direction_c;// the direction of the three axles, in radian
@@ -152,12 +153,12 @@ private:
 	vec flb_force;
 	vec flb_moment;
 	vec mres;            // resistence moment provided by normal distribution force
-	long double density; // specific gravity
-	long double mass;
-	long double volume;
+	REAL density; // specific gravity
+	REAL mass;
+	REAL volume;
 	vec  J;              // moment of inertia in local body-fixed frame
-	long double coef[10];// record particle's coefficients in global coordinates
-	long double kinetEnergy; // kinetic energy
+	REAL coef[10];// record particle's coefficients in global coordinates
+	REAL kinetEnergy; // kinetic energy
 	int  cntnum;
 
 public:
