+++ src/elements/DEM_ellip3d/ellip3d/shape.h
@@ -1,6 +1,7 @@
 #ifndef SHAPE_H
 #define SHAPE_H
 
+#include "realtypes.h"
 #include "vec.h"
 
 namespace dem {
@@ -8,7 +9,7 @@ namespace dem {
 class shape{ // abstract class
 public:
     virtual vec get_center() const =0;
-    virtual long double get_volume() const =0;
+    virtual REAL get_volume() const =0;
     virtual vec randomPoint() const =0;
     virtual void  print() const =0;
     virtual        ~shape() {}; // base class needs a virtual destructor.
